+++ HaruhiChokuretsuEditor/ControlExtensions.cs
@@ -1,4 +1,4 @@
-﻿using HaruhiChokuretsuLib;
+﻿using HaruhiChokuretsuLib.Archive;
 using System.Windows.Controls;
 
 namespace HaruhiChokuretsuEditor
