+++ HaruhiChokuretsuLib/Archive/DataFile.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HaruhiChokuretsuLib
+namespace HaruhiChokuretsuLib.Archive
 {
     public class DataFile : FileInArchive
     {
